backend/requirements.txt
# python version >= 3.11
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
httpx==0.24.0
google-generativeai>=0.3.0